requirements.txt
azure-identity
azure-mgmt-resourcegraph
msgraph-sdk
kuzu
streamlit
pyvis
pandas
python-dotenv
